% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI.R
\name{CI}
\alias{CI}
\title{Normal Confidence Interval}
\usage{
CI(x = 0, s = 1, n = 1, level = 0.95)
}
\arguments{
\item{x}{sample mean}

\item{s}{standard deviation}

\item{n}{sample size}

\item{level}{confidence level}
}
\value{
vector with two values containing the confidence interval for the
sample mean
}
\description{
Confidence interval for a normally distributed sample mean
}
\examples{
CI()
CI(150,5,30,.9)
}
