% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord2D.R
\name{coord2D}
\alias{coord2D}
\title{Plot a 2D Coordinate System}
\usage{
coord2D(x = 5, y = 5)
}
\arguments{
\item{x}{Distance from the origin to the maximum x-value.}

\item{y}{Distance from the origin to the maximum y-value.}
}
\value{
No return value, called for side effects
}
\description{
Plot a coordinate system in 2D with the origin in the center.
}
\examples{
coord2D()
}
