% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord3D.R
\name{coord3D}
\alias{coord3D}
\title{Plot a 3D Coordinate System}
\usage{
coord3D(th = 0, x = 10, y = 10, z = 10)
}
\arguments{
\item{th}{The angle at which the 3D plot should be displayed.}

\item{x}{Distance from the origin to the maximum x-value.}

\item{y}{Distance from the origin to the maximum y-value.}

\item{z}{Distance from the origin to the maximum z-value.}
}
\value{
A matrix containing the plot coordinates (used when adding features).
}
\description{
Plot a coordinate system in 3D with the origin bottom left.
}
\examples{
coord3D()
}
