% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSet.R
\name{dataSet}
\alias{dataSet}
\title{Obtain a Dataset from a Formula}
\usage{
dataSet(formula, data, subset = NULL)
}
\arguments{
\item{formula}{A formula}

\item{data}{A dataset}

\item{subset}{Either a logical vector or a vector of indices of the rows
to be returned. If NULL (default), all rows are returned.}
}
\value{
The dataset in \code{data} as a data table with variables as 
specified in \code{formula} and rows as specified by \code{subset}.
}
\description{
Given a formula, a dataset and a subset, retrieve the dataset that fulfills
the formula and subset.
}
\examples{
dataSet(mpg~.-disp,mtcars,10:20)
}
