% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isInt.R
\name{isInt}
\alias{isInt}
\title{Determine if the Input contains Integers}
\usage{
isInt(x, inf = TRUE)
}
\arguments{
\item{x}{integer or numeric type vector}

\item{inf}{logical field answering whether an infinite value should be 
considered an integer (default TRUE)}
}
\value{
TRUE for each value in \code{x} that is an integer, FALSE otherwise
}
\description{
Determine if numbers in a vector are integers (not just of integer type)
}
\examples{
isInt(c(3,3.23,Inf))
}
