% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linAlg.R
\name{laCrossProd}
\alias{laCrossProd}
\title{Cross Product (Linear Algebra)}
\usage{
laCrossProd(x, y)
}
\arguments{
\item{x}{vector of length 3.}

\item{y}{vector of length 3.}
}
\value{
Cross product of \code{x} and \code{y}.
}
\description{
Calculate the cross product as defined in linear algebra; note that this
differs from the cross product as defined by R.
}
\examples{
x=c(1,2,1)
y=1:3
laCrossProd(x,y)
}
