% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord3D.R
\name{lines3D}
\alias{lines3D}
\title{Lines in 3D}
\usage{
lines3D(pl, x, y, z, ...)
}
\arguments{
\item{pl}{Matrix containing the current plot coordinates.}

\item{x}{Vector with x-coordinates.}

\item{y}{Vector with y-coordinates.}

\item{z}{Vector with z-coordinates.}

\item{...}{additional graphical parameters (see lines()).}
}
\value{
No return value, called for side effects
}
\description{
Plot a line in a 3D plot through a set of points
}
\examples{
pl=coord3D(30)
lines3D(pl,0:10,0:10,rep(0,11))
lines3D(pl,0:10,0:10,c(0,2,1,3:8,7,5),col=2)
}
