% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmPartReg.R
\name{lmPartReg}
\alias{lmPartReg}
\title{Partial Regression Plot}
\usage{
lmPartReg(mod, pred, ...)
}
\arguments{
\item{mod}{A linear model object (obtained via the lm function)}

\item{pred}{The name (in quotes) of the predictor for which the plot should
be produced}

\item{...}{Any other arguments to be passed to the plot}
}
\value{
A partial regression plot for \code{pred} in the linear model 
\code{mod}
}
\description{
Plot the partial regression plot for one of the predictors of a linear model
}
\examples{
lmod=lm(mpg~.,mtcars)
lmPartReg(lmod,"wt")
}
