% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\title{Find Outliers}
\usage{
outliers(x)
}
\arguments{
\item{x}{vector}
}
\value{
A list with a variable \code{idx} containing the indices of the
outliers and a variable \code{values} containing the values of the outliers.
}
\description{
Find the outliers in a vector of values.
}
\examples{
x=c(100,30:40,101,25:28)
outliers(x)
}
