% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCol.R
\name{plotCol}
\alias{plotCol}
\title{Plot Colors}
\usage{
plotCol(col)
}
\arguments{
\item{col}{vector with colors}
}
\value{
A plot showing the colors in \code{col}
}
\description{
Plot one or more colors
}
\examples{
plotCol("maroon")
}
