% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popVarSd.R
\name{pop.sd}
\alias{pop.sd}
\title{Calculate the Population Standard Deviation}
\usage{
pop.sd(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
The population standard deviation of the entries in \code{x}
}
\description{
Calculate the standard deviation of a numeric vector if the data constitutes
the whole population. Note that missing values are excluded.
}
\examples{
pop.sd(c(1:6,NA,7:10))
}
