% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popVarSd.R
\name{pop.var}
\alias{pop.var}
\title{Calculate the Population Variance}
\usage{
pop.var(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
The population variance of the entries in \code{x}
}
\description{
Calculate the variance of a numeric vector if the data constitutes the
whole population. Note that missing values are excluded.
}
\examples{
pop.var(c(1:6,NA,7:10))
}
