% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqlineHalf.R
\name{qqlineHalf}
\alias{qqlineHalf}
\title{Line through a Half-Normal Plot}
\usage{
qqlineHalf(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
No return value, called for side effects
}
\description{
Plot a line through the first and third quantile of a halfnormal line
}
\examples{
z=rnorm(100)
faraway::halfnorm(z)
qqlineHalf(z)
}
