% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statLearn.R
\name{ROCknn}
\alias{ROCknn}
\title{KNN ROC curve}
\usage{
ROCknn(mod, response)
}
\arguments{
\item{mod}{The output of the knn function, run with prob=TRUE}

\item{response}{A vector with responses for the testing dataset used to run
the knn function.}
}
\value{
A plot with the ROC curve will be produced, nothing is returned.
}
\description{
Plot the ROC curve for a KNN model. Note that it can only be used when
the response is dichotomous.
}
\examples{
yhat=class::knn(Puromycin[,c("conc","rate")],Puromycin[,c("conc","rate")],
         Puromycin$state,10,prob=TRUE)
ROCknn(yhat,Puromycin$state)
}
