% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round2.R
\name{round2}
\alias{round2}
\title{Round to the Nearest Number}
\usage{
round2(x, digits = 0)
}
\arguments{
\item{x}{number to be rounded}

\item{digits}{number of digits to round to}
}
\value{
Number rounded to the number of \code{digits} indicated
}
\description{
Round to the nearest number with the number of digits as indicated. NOTE:
Unlike the base \code{round} function it rounds a 5 to the higher number,
rather than the nearest even number.
}
\examples{
round2(2.5)
}
