\name{smallstuff-package}
\alias{smallstuff-package}
\alias{smallstuff}
\docType{package}
\title{
  Dr. Small's Data Science Functions
}
\description{
  Functions used by students in the Master's of Data Science program at Drew
  University. 
}
\details{
  Some functions are used for Statistics using R, such as pop.var (calculates
  the population variance), and outliers (finds the outliers in a distribution
  with their indices), some for Applied Regression Analysis such as projMatrix
  (Calculates the projection matrix) and systemEq (solves a system of linear 
  equations), some for Machine Learning such as lmSub (finds the best linear
  model in subset selection), and some for Networks such as get_subgraphs, which
  splits a graph into subgraphs.
}
\author{
Ellie Small, esmall1@drew.edu.

Maintainer: Ellie Small <esmall1@drew.edu>
}
