% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord3D.R
\name{span3D}
\alias{span3D}
\title{Span of a Matrix}
\usage{
span3D(M, th = 0, ph = 15)
}
\arguments{
\item{M}{Matrix for which the span should be shown.}

\item{th}{A vector indicating the horizontal angle at which the plot should
be shown.}

\item{ph}{A vector indicating the vertical angle at which the plot should be
shown.}
}
\value{
A matrix containing the plot coordinates (used when adding features).
}
\description{
Displays a perspective plot showing the plane that is the span of a matrix
}
\examples{
span3D(matrix(c(1,0,0,1,1,1),3))
}
