% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systemEq.R
\name{systemEq}
\alias{systemEq}
\title{Solve a System of Equations}
\usage{
systemEq(A, y)
}
\arguments{
\item{A}{matrix A in Ax=y}

\item{y}{output vector in Ax=y}
}
\value{
the unique solution x to Ax=y
}
\description{
Solve a system of equations if it has a unique solution;
output an error message otherwise
}
\examples{
systemEq(matrix(c(1:3,2,4,4),3),c(3,6,7))
}
