% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord2D.R
\name{vector2D}
\alias{vector2D}
\title{Add a Vector to a 2D Coordinate System}
\usage{
vector2D(v, fr = c(0, 0), col = 2)
}
\arguments{
\item{v}{A vector with 2 entries.}

\item{fr}{Vector containing the point at which the vector should start
(defaults to the origin).}

\item{col}{Color of the vector (defaults to red).}
}
\value{
No return value, called for side effects
}
\description{
Add a Vector to a 2D Coordinate System
}
\examples{
a=c(2,4)
b=c(0,3)
coord2D()
vector2D(a)
vector2D(b)
vector2D(a-b,b,"blue")
}
