% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord3D.R
\name{vector3D}
\alias{vector3D}
\title{Add a Vector to a 3D Coordinate System}
\usage{
vector3D(pl, v, fr = rep(0, 3), col = "red")
}
\arguments{
\item{pl}{Matrix containing the current plot coordinates.}

\item{v}{A vector with 3 entries.}

\item{fr}{The point at which the vector should start
(defaults to the origin).}

\item{col}{Color of the vector (defaults to red).}
}
\value{
No return value, called for side effects
}
\description{
Add a Vector to a 3D Coordinate System
}
\examples{
a=c(2,4,8)
b=c(6,0,4)
pl=coord3D()
vector3D(pl,a)
vector3D(pl,b)
vector3D(pl,a-b,b,3)
}
