% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withinPC.R
\name{withinPC}
\alias{withinPC}
\title{Calculate Row or Column Percentages}
\usage{
withinPC(X, rows = TRUE, rnd = 1)
}
\arguments{
\item{X}{matrix or table}

\item{rows}{TRUE (default) to calculate by rows, or FALSE to calculate by
columns}

\item{rnd}{numbers of digits to round the result to}
}
\value{
A matrix or table with percentages
}
\description{
Calculate percentages of values in a matrix or table with respect to the
row or column totals.
}
\examples{
(X=matrix(c(1:12),3))
withinPC(X)
}
